refactor: App.js에 있던 user 리스트 렌더링하는 로직 Users 컴포넌트로 분리
